+++ constants.env
@@ -35,5 +35,6 @@ CARGO_UDEPS_VERSION=0.1.60
 CARGO_WORKSPACES_VERSION=0.4.2
 
 # Other tools
+GUNGRAUN_RUNNER_VERSION=0.19.0
 JUST_VERSION=1.46.0
 SCCACHE_VERSION=v0.13.0
