+++ crates/multitude/Cargo.toml
@@ -47,9 +47,6 @@ bytes = ["dep:bytes"]
 bytemuck = ["dep:bytemuck", "dst"]
 bytesbuf = ["dep:bytesbuf", "std"]
 zerocopy = ["dep:zerocopy", "dst"]
-# Internal feature: gates gungraun (Valgrind-based) benchmarks that require
-# callgrind to be installed. Not intended for external use.
-_gungraun = []
 
 [dependencies]
 allocator-api2 = { workspace = true, features = ["alloc"] }
@@ -74,9 +71,14 @@ bolero = { workspace = true, features = ["std"] }
 # but does not itself activate that feature on `bolero-engine`.
 bumpalo = { workspace = true, features = ["collections"] }
 criterion.workspace = true
-gungraun = { workspace = true, features = ["default"] }
 mutants.workspace = true
 serde_json.workspace = true
+
+# Workspace dep declares `default-features = false`; we re-enable `default` here
+# (which pulls in the `benchmark` feature with `library_benchmark`,
+# `library_benchmark_group!`, and `gungraun::main!`) only on Linux.
+[target.'cfg(target_os = "linux")'.dev-dependencies]
+gungraun = { workspace = true, features = ["default"] }
 
 [lints]
 workspace = true
@@ -89,15 +91,16 @@ harness = false
 name = "criterion_drop"
 harness = false
 
+# Callgrind benches require Linux (Valgrind). The bench files are gated to compile
+# to a no-op on non-Linux targets, but the [[bench]] entry itself cannot be
+# cfg-gated, so it is unconditional here.
 [[bench]]
 name = "gungraun_alloc"
 harness = false
-required-features = ["_gungraun"]
 
 [[bench]]
 name = "gungraun_drop"
 harness = false
-required-features = ["_gungraun"]
 
 [[example]]
 name = "basic"
